pyproject.toml
[build-system]
requires = ["setuptools>=77", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "gcn-kafka"
description = "Kafka client for NASA's General Coordinates Network (GCN)"
license = "CC0-1.0"
readme = "README.md"
authors = [
    { name = "Leo Singer", email = "leo.p.singer@nasa.gov" },
    { name = "Tom Barclay", email = "tb@umbc.edu" },
    { name = "Eric Burns", email = "ericburns@lsu.edu" }
]
classifiers = [
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: System :: Networking",
    "Typing :: Typed",
]
dependencies = [
    "authlib",
    "certifi",
    "confluent-kafka >= 2.2.0",
    "requests",
    "pyjwt"
]
requires-python = ">=3.9"
dynamic = [ "version" ]

[project.urls]
source = "https://github.com/nasa-gcn/gcn-kafka-python"

[project.optional-dependencies]
test = ["pytest"]

[tool.setuptools_scm]
write_to = "src/gcn_kafka/_version.py"

[tool.coverage.run]
source = ["test"]
source_pkgs = ["gcn_kafka"]
